refactor - Monster Class
- Monster 클래스 추가 및 기존 몬스터 클래스 공통 로직 이동.
- 중복 코드 제거 및 std::max를 활용한 로직 간소화.
- 새로운 Monster.cpp/h 파일로 관련 코드 분리.

+++ TextRPG/System/GameSession.cpp
@@ -84,7 +84,7 @@ void GameSession::visitShop()
             Item* item = player->getInventory().getItem(sellChoice - 1);
             if (item != nullptr)
             {
-                int sellPrice = item->getPrice() * 0.6;
+                int sellPrice = static_cast<int>(item->getPrice() * 0.6);
                 gold += sellPrice;
                 shop.addItem(item->clone());
